requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
python-dotenv==1.0.1
flasgger==0.9.7.1
boto3==1.34.162
python-dateutil==2.9.0.post0
flask-migrate==4.0.5
alembic==1.13.1
PyJWT==2.8.0
Werkzeug==3.0.3
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
